Update and refactor tests
