+++ tests/test_hssm.py
@@ -14,49 +14,6 @@ from hssm.likelihoods import DDM, logp_ddm
 pytensor.config.floatX = "float32"
 
 
-@pytest.fixture
-def data():
-    v_true, a_true, z_true, t_true = [0.5, 1.5, 0.5, 1.5]
-    obs_ddm = ssms.basic_simulators.simulator(
-        [v_true, a_true, z_true, t_true], model="ddm", n_samples=50
-    )
-    obs_ddm = np.column_stack([obs_ddm["rts"][:, 0], obs_ddm["choices"][:, 0]])
-    dataset = pd.DataFrame(obs_ddm, columns=["rt", "response"])
-    dataset["x"] = dataset["rt"] * 0.1
-    dataset["y"] = dataset["rt"] * 0.5
-    return dataset
-
-
-@pytest.fixture(scope="module")
-def fixture_path():
-    return Path(__file__).parent / "fixtures"
-
-
-@pytest.fixture
-def data_angle():
-    v_true, a_true, z_true, t_true, theta_true = [0.5, 1.5, 0.5, 0.5, 0.3]
-    obs_angle = ssms.basic_simulators.simulator(
-        [v_true, a_true, z_true, t_true, theta_true], model="angle", n_samples=100
-    )
-    obs_angle = np.column_stack([obs_angle["rts"][:, 0], obs_angle["choices"][:, 0]])
-    data = pd.DataFrame(obs_angle, columns=["rt", "response"])
-    return data
-
-
-@pytest.fixture
-def example_model_config():
-    return {
-        "loglik_kind": "example",
-        "list_params": ["v", "a", "z", "t"],
-        "bounds": {
-            "v": (-3.0, 3.0),
-            "a": (0.3, 2.5),
-            "z": (0.1, 0.9),
-            "t": (0.0, 2.0),
-        },
-    }
-
-
 @pytest.mark.parametrize(
     "include, should_raise_exception",
     [
@@ -129,12 +86,12 @@ def example_model_config():
         ),
     ],
 )
-def test_transform_params_general(data, include, should_raise_exception):
+def test_transform_params_general(data_ddm_reg, include, should_raise_exception):
     if should_raise_exception:
         with pytest.raises(Exception):
-            HSSM(data=data, include=include)
+            HSSM(data=data_ddm_reg, include=include)
     else:
-        model = HSSM(data=data, include=include)
+        model = HSSM(data=data_ddm_reg, include=include)
         # Check model properties using a loop
         param_names = ["v", "a", "z", "t", "p_outlier"]
         model_param_names = list(model.params.keys())
@@ -142,28 +99,30 @@ def test_transform_params_general(data, include, should_raise_exception):
         assert len(model.params) == 5
 
 
-def test_custom_model(data, example_model_config):
+def test_custom_model(data_ddm):
     with pytest.raises(
         ValueError, match="When using a custom model, please provide a `loglik_kind.`"
     ):
-        model = HSSM(data=data, model="custom")
+        model = HSSM(data=data_ddm, model="custom")
 
     with pytest.raises(
         ValueError, match="Please provide `list_params` via `model_config`."
     ):
-        model = HSSM(data=data, model="custom", loglik_kind="analytical")
+        model = HSSM(data=data_ddm, model="custom", loglik_kind="analytical")
 
     with pytest.raises(
         ValueError, match="Please provide `list_params` via `model_config`."
     ):
-        model = HSSM(data=data, model="custom", loglik=DDM, loglik_kind="analytical")
+        model = HSSM(
+            data=data_ddm, model="custom", loglik=DDM, loglik_kind="analytical"
+        )
 
     with pytest.raises(
         ValueError,
         match="Please provide `list_params` via `model_config`.",
     ):
         model = HSSM(
-            data=data,
+            data=data_ddm,
             model="custom",
             loglik=DDM,
             loglik_kind="analytical",
@@ -171,7 +130,7 @@ def test_custom_model(data, example_model_config):
         )
 
     model = HSSM(
-        data=data,
+        data=data_ddm,
         model="custom",
         model_config={
             "list_params": ["v", "a", "z", "t"],
@@ -191,14 +150,14 @@ def test_custom_model(data, example_model_config):
     assert model.list_params == ["v", "a", "z", "t", "p_outlier"]
 
 
-def test_model_definition_outside_include(data):
-    model_with_one_param_fixed = HSSM(data, a=0.5)
+def test_model_definition_outside_include(data_ddm):
+    model_with_one_param_fixed = HSSM(data_ddm, a=0.5)
 
     assert "a" in model_with_one_param_fixed.priors
     assert model_with_one_param_fixed.priors["a"] == 0.5
 
     model_with_one_param = HSSM(
-        data, a={"prior": {"name": "Normal", "mu": 0.5, "sigma": 0.1}}
+        data_ddm, a={"prior": {"name": "Normal", "mu": 0.5, "sigma": 0.1}}
     )
 
     assert "a" in model_with_one_param.priors
@@ -207,7 +166,7 @@ def test_model_definition_outside_include(data):
     with pytest.raises(
         ValueError, match='Parameter "a" is already specified in `include`'
     ):
-        HSSM(data, include=[{"name": "a", "prior": 0.5}], a=0.5)
+        HSSM(data_ddm, include=[{"name": "a", "prior": 0.5}], a=0.5)
 
 
 def test_model_with_approx_differentiable_likelihood_type(data_angle):
@@ -217,10 +176,10 @@ def test_model_with_approx_differentiable_likelihood_type(data_angle):
     assert model.loglik == download_hf(loglik)
 
 
-def test_sample_prior_predictive(data):
-    data = data.iloc[:10, :]
+def test_sample_prior_predictive(data_ddm_reg):
+    data_ddm_reg = data_ddm_reg.iloc[:10, :]
 
-    model_no_regression = HSSM(data=data)
+    model_no_regression = HSSM(data=data_ddm_reg)
     rng = np.random.default_rng()
 
     prior_predictive_1 = model_no_regression.sample_prior_predictive(draws=10)
@@ -228,14 +187,18 @@ def test_sample_prior_predictive(data):
         draws=10, random_seed=rng
     )
 
-    model_regression = HSSM(data=data, include=[dict(name="v", formula="v ~ 1 + x")])
+    model_regression = HSSM(
+        data=data_ddm_reg, include=[dict(name="v", formula="v ~ 1 + x")]
+    )
     prior_predictive_3 = model_regression.sample_prior_predictive(draws=10)
 
-    model_regression_a = HSSM(data=data, include=[dict(name="a", formula="a ~ 1 + x")])
+    model_regression_a = HSSM(
+        data=data_ddm_reg, include=[dict(name="a", formula="a ~ 1 + x")]
+    )
     prior_predictive_4 = model_regression_a.sample_prior_predictive(draws=10)
 
     model_regression_multi = HSSM(
-        data=data,
+        data=data_ddm_reg,
         include=[
             dict(name="v", formula="v ~ 1 + x"),
             dict(name="a", formula="a ~ 1 + y"),
@@ -243,10 +206,10 @@ def test_sample_prior_predictive(data):
     )
     prior_predictive_5 = model_regression_multi.sample_prior_predictive(draws=10)
 
-    data["subject_id"] = np.arange(10)
+    data_ddm_reg["subject_id"] = np.arange(10)
 
     model_regression_random_effect = HSSM(
-        data=data,
+        data=data_ddm_reg,
         include=[
             dict(name="v", formula="v ~ (1|subject_id) + x"),
             dict(name="a", formula="a ~ (1|subject_id) + y"),
@@ -257,25 +220,25 @@ def test_sample_prior_predictive(data):
     )
 
 
-def test_hierarchical(data):
-    data = data.iloc[:10, :].copy()
-    data["participant_id"] = np.arange(10)
+def test_hierarchical(data_ddm):
+    data_ddm = data_ddm.iloc[:10, :].copy()
+    data_ddm["participant_id"] = np.arange(10)
 
-    model = HSSM(data=data)
+    model = HSSM(data=data_ddm)
     assert all(
         param.is_regression
         for name, param in model.params.items()
         if name != "p_outlier"
     )
 
-    model = HSSM(data=data, v=bmb.Prior("Uniform", lower=-10.0, upper=10.0))
+    model = HSSM(data=data_ddm, v=bmb.Prior("Uniform", lower=-10.0, upper=10.0))
     assert all(
         param.is_regression
         for name, param in model.params.items()
         if name not in ["v", "p_outlier"]
     )
 
-    model = HSSM(data=data, a=bmb.Prior("Uniform", lower=0.0, upper=10.0))
+    model = HSSM(data=data_ddm, a=bmb.Prior("Uniform", lower=0.0, upper=10.0))
     assert all(
         param.is_regression
         for name, param in model.params.items()
@@ -283,7 +246,7 @@ def test_hierarchical(data):
     )
 
     model = HSSM(
-        data=data,
+        data=data_ddm,
         v=bmb.Prior("Uniform", lower=-10.0, upper=10.0),
         a=bmb.Prior("Uniform", lower=0.0, upper=10.0),
     )
