+++ tests/test_likelihoods.py
@@ -4,31 +4,16 @@ This code compares WFPT likelihood function with
 old implementation of WFPT from (https://github.com/hddm-devs/hddm)
 """
 import math
-import random
 
 import numpy as np
 import pytest
-import ssms.basic_simulators
 from numpy.random import rand
 
 # pylint: disable=C0413
 from hssm.likelihoods.analytical import compare_k, logp_ddm_sdv
 
 
-@pytest.fixture
-def data_fixture():
-    random_integer = random.randint(0, 9)
-    v_true, a_true, z_true, t_true = [0.5, 1.5, 0.5, 0.5]
-    obs_ddm = ssms.basic_simulators.simulator(
-        [v_true, a_true, z_true, t_true],
-        model="ddm",
-        n_samples=1000,
-        random_state=random_integer,
-    )
-    return np.column_stack([obs_ddm["rts"][:, 0], obs_ddm["choices"][:, 0]])
-
-
-def test_kterm(data_fixture):
+def test_kterm(data_ddm):
     """This function defines a range of kterms and tests results to
     makes sure they are not equal to infinity or unknown values.
     """
@@ -39,19 +24,19 @@ def test_kterm(data_fixture):
         z = 0.5 * rand()
         t = rand() * 0.5
         err = 1e-7
-        logp = logp_ddm_sdv(data_fixture, v, a, z, t, sv, err, k_terms=k_term)
+        logp = logp_ddm_sdv(data_ddm, v, a, z, t, sv, err, k_terms=k_term)
         logp = sum(logp.eval())
         assert not math.isinf(logp)
         assert not math.isnan(logp)
 
 
-def test_compare_k(data_fixture):
+def test_compare_k(data_ddm):
     """This function tests output of decision function."""
     err = 1e-7
-    data = data_fixture[:, 0] * data_fixture[:, 1]
-    lambda_rt = compare_k(np.abs(data), err)
+    data = data_ddm["rt"] * data_ddm["response"]
+    lambda_rt = compare_k(np.abs(data.values), err)
     assert all(not v for v in lambda_rt.eval())
-    assert data_fixture.shape[0] == lambda_rt.eval().shape[0]
+    assert data_ddm.shape[0] == lambda_rt.eval().shape[0]
 
 
 # def test_logp(data_fixture):
@@ -87,28 +72,28 @@ def shared_params():
     }
 
 
-def test_no_inf_values_a(data_fixture, shared_params):
+def test_no_inf_values_a(data_ddm, shared_params):
     for a in np.arange(2.5, 5.1, 0.5):
         params = {**shared_params, "a": a}
-        logp = logp_ddm_sdv(data_fixture, **params)
+        logp = logp_ddm_sdv(data_ddm, **params)
         assert np.all(
             np.isfinite(logp.eval())
         ), f"log_pdf_sv() returned non-finite values for a = {a}."
 
 
-def test_no_inf_values_t(data_fixture, shared_params):
+def test_no_inf_values_t(data_ddm, shared_params):
     for t in np.arange(3.0, 5.1, 0.5):
         params = {**shared_params, "t": t}
-        logp = logp_ddm_sdv(data_fixture, **params)
+        logp = logp_ddm_sdv(data_ddm, **params)
         assert np.all(
             np.isfinite(logp.eval())
         ), f"log_pdf_sv() returned non-finite values for t = {t}."
 
 
-def test_no_inf_values_v(data_fixture, shared_params):
+def test_no_inf_values_v(data_ddm, shared_params):
     for v in np.arange(3.0, 5.1, 0.5):
         params = {**shared_params, "v": v}
-        logp = logp_ddm_sdv(data_fixture, **params)
+        logp = logp_ddm_sdv(data_ddm, **params)
         assert np.all(
             np.isfinite(logp.eval())
         ), f"log_pdf_sv() returned non-finite values for v = {v}."
